+++ eInspektor/InspectorView.cs
@@ -21,8 +21,13 @@ namespace eInspektor
         
         private void nazadToolStripMenuItem_Click(object sender, EventArgs e)
         {
-            this.Close();
-            startForm.Show();
+            DialogResult result = MessageBox.Show("Izmjene nisu sačuvane. Želite li ih sačuvati prije nego što napustite rad sa inspektorima?","Inspektori", MessageBoxButtons.OKCancel, MessageBoxIcon.Question);
+            if (result == DialogResult.OK) {
+                //sacuvaj
+                this.Close();
+                startForm.Show();
+            }
+           
         }
 
         private void InspectorView_Load(object sender, EventArgs e)
@@ -67,27 +72,31 @@ namespace eInspektor
 
         private void obrišiToolStripMenuItem_Click(object sender, EventArgs e)
         {
-            foreach (var cell in dataGridView1.SelectedCells)
-            {
-                int id = 0;
-                try
+            DialogResult result = MessageBox.Show("Da li ste sigurni da zelite obrisati odabranog inspektora?", "Brisanje inspektora", MessageBoxButtons.OKCancel, MessageBoxIcon.Question);
+            if (result == DialogResult.OK) {
+                foreach (var cell in dataGridView1.SelectedCells)
                 {
-                    id = (int)dataGridView1.CurrentCell.OwningRow.Cells[0].Value;
-                }
-                catch (NullReferenceException)
-                {
-                    continue;
-                }
+                    int id = 0;
+                    try
+                    {
+                        id = (int)dataGridView1.CurrentCell.OwningRow.Cells[0].Value;
+                    }
+                    catch (NullReferenceException)
+                    {
+                        continue;
+                    }
 
-                inspector i = db.inspectors.Find(id);
-                if (i == null)
-                {
-                    continue;
+                    inspector i = db.inspectors.Find(id);
+                    if (i == null)
+                    {
+                        continue;
+                    }
+                    db.inspectors.Remove(i);
                 }
-                db.inspectors.Remove(i);
+                db.SaveChanges();
+                InspectorView_Load(sender, e);
             }
-            db.SaveChanges();
-            InspectorView_Load(sender, e);
+           
         }
     }
 }
